requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
